New tactics bfs/dfs eauto.
The two are based on the same implementation as typeclasses eauto,
although they try to be as compatible as possible with the current eauto.
The only known difference is wrt to the search-depth definition:
these new tactics use the same search-depth definition as in auto
(limit the depth of the proof-search tree) instead of limiting the size
of the proof-search tree like eauto does. See bug 5555.